LayoutTests/platform/chromium/fast/js/parse-error-external-script-in-eval-expected.txt

CONSOLE MESSAGE: line 3: Uncaught SyntaxError: Unexpected identifier
This tests that a line number and error message is printed for a parse error in a separate source file using eval().
